requirements.txt
flask==3.0.2
PyPDF2==3.0.1
python-dotenv==1.0.1
openai==0.28.1
python-magic==0.4.27
Werkzeug==3.0.1
flask-cors==4.0.0 